ui/api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-multipart>=0.0.6
pytest>=8.0.0
pytest-asyncio>=0.23.0
psycopg2-binary>=2.9.0
